# docs/service/catalog-service/api.yaml
openapi: 3.0.3
info:
  title: catalog-service API
  version: "0.1"
  description: |
    Quản lý Catalog — Brand, Category, AttributeTemplate, Product (SPU), Variant (SKU).
    Upstream thuần túy: không consume event, chỉ publish.

    **Auth:**
    - `/api/**` — public (không cần token)
    - `/api/admin/**` — yêu cầu role `ADMIN` (bearer token từ oauth2-service)
servers:
  - url: http://localhost:8082
    description: Local dev
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          example: null

    # ── Brand ─────────────────────────────────────────────────────────────────
    BrandResponse:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          description: ULID
          example: "01HX4KZ7B3QNMR2VFPJ6W8Y9T"
        name:
          type: string
          example: "Nike"
        slug:
          type: string
          example: "nike"
    BrandPage:
      type: object
      required: [items, total]
      description: Danh sách công khai — chỉ ACTIVE.
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BrandResponse' }
        total: { type: integer, format: int64 }
    BrandAdminSummaryPage:
      type: object
      required: [items, total]
      description: Danh sách quản trị — mọi trạng thái.
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, name, slug, status]
            properties:
              id: { type: string }
              name: { type: string }
              slug: { type: string }
              status: { $ref: '#/components/schemas/BrandStatus' }
        total: { type: integer, format: int64 }
    BrandDetailResponse:
      type: object
      required: [id, name, slug, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        status: { $ref: '#/components/schemas/BrandStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BrandStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    BrandOptionResponse:
      type: object
      required: [options]
      description: Droplist cho seller khi khai brand cho sản phẩm.
      properties:
        options:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id: { type: string }
              name: { type: string }
    CreateBrandRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          example: "Nike"
        slug:
          type: string
          example: "nike"
    UpdateBrandRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Nike International"

    # ── Category ──────────────────────────────────────────────────────────────
...
  responses:
    400:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "VALIDATION_ERROR" }
                  message: { type: string }
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    409:
      description: Conflict (duplicate, in-use)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
...
paths:
...
  /api/brands:
    get:
      tags: [Brand (Public)]
      summary: Danh sách brand active — tìm + phân trang
      description: Không cần auth. Đọc DB trực tiếp (không cache — xem data.md § Cache).
      operationId: listBrands
      parameters:
        - { name: keyword, in: query, required: false, schema: { type: string }, description: "Tìm trong name, không phân biệt hoa/thường" }
        - { name: page,    in: query, required: false, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: size,    in: query, required: false, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BrandPage'

  # ── Brand (Seller) ───────────────────────────────────────────────────────────
  /api/seller/brands:
    get:
      tags: [Brand (Seller)]
      summary: Droplist brand khi khai sản phẩm
      description: Chỉ ACTIVE, gõ tới đâu gợi ý tới đó, tối đa 20 kết quả. Không phân trang (không cần cho autocomplete).
      operationId: searchBrandsForSeller
      parameters:
        - { name: keyword, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BrandOptionResponse'

  # ── Brand (Admin) ──────────────────────────────────────────────────────────
  /api/admin/brands:
    get:
      tags: [Brand (Admin)]
      summary: Danh sách brand (quản trị) — mọi trạng thái, lọc + phân trang
      operationId: listAdminBrands
      parameters:
        - { name: keyword, in: query, required: false, schema: { type: string } }
        - { name: status,  in: query, required: false, schema: { $ref: '#/components/schemas/BrandStatus' } }
        - { name: page,    in: query, required: false, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: size,    in: query, required: false, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BrandAdminSummaryPage'
    post:
      tags: [Brand (Admin)]
      summary: Tạo brand mới
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBrandRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BrandResponse'
        "400":
          $ref: '#/components/responses/400'
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: BRAND_NAME_ALREADY_EXISTS | BRAND_SLUG_ALREADY_EXISTS
  /api/admin/brands/{id}:
    get:
      tags: [Brand (Admin)]
      summary: Chi tiết brand (quản trị)
      operationId: getBrand
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BrandDetailResponse'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: BRAND_NOT_FOUND
    put:
      tags: [Brand (Admin)]
      summary: Cập nhật tên brand
      description: Chỉ cho phép đổi `name`. `slug` là immutable sau khi tạo.
      operationId: updateBrand
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Brand ULID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBrandRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: BRAND_NOT_FOUND
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: BRAND_NAME_ALREADY_EXISTS
    delete:
      tags: [Brand (Admin)]
      summary: Vô hiệu hóa brand (soft toggle, không phải xoá dữ liệu)
      description: |
        Chuyển brand sang INACTIVE — ẩn khỏi `ListBrands`/storefront/droplist seller. KHÔNG guard theo
        "đang có Product dùng hay không" (đã bỏ guard `BRAND_IN_USE` — chặn ngược với nguyên tắc soft-toggle,
        xem service.md § Category lifecycle). Product cũ tham chiếu brand đã deactivate vẫn resolve bình
        thường qua `GetProduct` (không lọc theo status).
      operationId: deactivateBrand
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
  /api/admin/brands/{id}/activate:
    post:
      tags: [Brand (Admin)]
      summary: Kích hoạt lại brand đã tắt
      operationId: activateBrand
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: BRAND_NOT_FOUND

  # ── AttributeTemplate (Admin) ───────────────────────────────────────────────
...